requirements.txt
backports-abc==0.5
beautifulsoup4==4.6.0
bs4==0.0.1
certifi==2017.11.5
chardet==3.0.4
idna==2.6
Naked==0.1.31
pbr==3.1.1
pycrypto==2.6.1
PyExecJS==1.5.0
PyYAML==3.12
requests==2.18.4
shellescape==3.4.1
singledispatch==3.4.0.3
six==1.11.0
stevedore==1.27.1
tornado==4.5.2
urllib3==1.22
virtualenv==15.1.0
virtualenv-clone==0.2.6
virtualenvwrapper==4.8.2
